kenechi_ojukwu/requirements.txt
deep-sort-realtime==1.3.2
facenet-pytorch==2.5.2
fire==0.5.0
pytube==12.1.3